# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VR GARLANDS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google:
    post:
      operationId: googleLogin
      tags: [auth]
      summary: Login with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get user cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List user orders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create Razorpay order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentOrderRequest"
      responses:
        "200":
          description: Razorpay order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify Razorpay payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GoogleLoginRequest:
      type: object
      properties:
        idToken:
          type: string
          description: Google ID token from OAuth flow
      required:
        - idToken
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatar:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - createdAt
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        discountedPrice:
          type: number
          nullable: true
        unit:
          type: string
        image:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: integer
          description: Delivery time in minutes
        discountPercent:
          type: integer
          nullable: true
      required:
        - id
        - name
        - price
        - unit
        - image
        - categoryId
        - categoryName
        - inStock
        - isFeatured
        - rating
        - reviewCount
        - deliveryTime
...
    CartItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - productId
        - quantity
        - product
    Cart:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required:
        - id
        - items
        - total
        - itemCount
    AddCartItemRequest:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          default: 1
      required:
        - productId
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
          minimum: 0
      required:
        - quantity
...
    Address:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - name
        - phone
        - address
        - city
        - state
        - pincode
        - isDefault
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        image:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - name
        - image
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        address:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
      required:
        - id
        - status
        - paymentStatus
        - total
        - items
        - address
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        addressId:
          type: string
        couponCode:
          type: string
      required:
        - addressId
    CreatePaymentOrderRequest:
      type: object
      properties:
        orderId:
          type: string
      required:
        - orderId
    PaymentOrderResponse:
      type: object
      properties:
        razorpayOrderId:
          type: string
        amount:
          type: number
        currency:
          type: string
        keyId:
          type: string
      required:
        - razorpayOrderId
        - amount
        - currency
        - keyId
    VerifyPaymentRequest:
      type: object
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
        orderId:
          type: string
      required:
        - razorpayOrderId
        - razorpayPaymentId
        - razorpaySignature
        - orderId
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message